oracle_rag_eval_app/requirements.txt
streamlit==1.37.0
python-dotenv==1.0.1
oracledb==2.3.0
openai==1.50.2
tiktoken==0.7.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
matplotlib==3.9.0
sqlparse==0.5.1
sentence-transformers==3.0.1
ollama
